refactor(apps/official-website): change hooks file naming
