+++ apps/official-website/src/lib/hooks/index.ts
@@ -1,2 +1,3 @@
-export { default as useMedia } from './useMedia';
-export { default as useIsMobile } from './useIsMobile';
+export { default as useAcceptPattern } from './use-accept-pattern';
+export { default as useIsMobile } from './use-is-mobile';
+export { default as useMedia } from './use-media';
