--- apps/official-website/src/lib/hooks/useIsMobile.ts
@@ -1,12 +1,0 @@
-import { useMedia } from ".";
-
-const useIsMobile = () => {
-  const isMobileSize = useMedia("(max-width: 768px)");
-
-  return (
-    /Android|webOS|iPhone|iPad|iPod|BlackBerry/i.test(navigator.userAgent) ||
-    isMobileSize
-  );
-};
-
-export default useIsMobile;
